requirements.txt
flask
openai
faiss-cpu
langdetect
beautifulsoup4
requests
numpy
gunicorn
